net: Avoid using logging methods in context.T
This is something I need to figure out, any logging through the context
seems to cause trouble (hang the goroutine?) in Android. For now, just
get rid of the uses in triangle code for this logging. However, if a log
is triggered in the v23 implementation then the goroutine will get stuck
as well, so I should debug that and return to using context logging.

For now, just do what it takes to get the network stuff to work, albeit
in a fragile way, on android.