Discovery based connectivity.
Seems to work - devices find each other, invite each other and then
figure things out when a device vanishes.

However, as implemented we end up with a cyclic screen: If there are two
devices, each invites the other and we end up with a tight-loop - the
two screens are simultaneously to the left and right of each other
and thus form a strong clique with no incentive to invite a third
screen that might start advertising later on.