net: Stop using streaming RPCs.
Instead of a streaming RPC, use separate "Invite" and "Give" RPCs.
The main motivation is responsiveness: Doing so allows the caller
to get confirmation when the triangle has reached the other end.
With the previous streaming style, there was no confirmation as
to whether a single element in the stream reached the other end.
It seems the previous use was just not appropriate for streaming.

While at it, I think this simplifies the networking logic as well.
Complete state is now contained in one single goroutine/function
(netManager.run) and the core logic of what to do when a new
peer joins or leaves is easier to follow in that function.

Still haven't solved the "create a cycle" problem.